Android Application
Android application created to supply the user’s with an electronic
photo diary. The application allows the user to attach a photo from the
local device memory to a note. This adds a visual touch to the note
keeping experience. Also the app saves all the notes and gives access to
the user via an aesthetic and functional interface.